skills/rfp-to-proposal-pipeline/scripts/requirements.txt
pdfplumber>=0.11.0
google-genai>=0.8.0
python-dotenv>=1.0.1

# analyze_reference.py (reference palette extraction)
Pillow>=10.0
scikit-learn>=1.3
numpy>=1.26
requests>=2.31
